now work

+++ First Internal Trainer/Graphical.h
@@ -6,11 +6,20 @@ class Graphical {
 		
 		/*
 		* Return Values
-		* 0 = false
-		* 1 = true
-		* 
+		* 0 = No
+		* 1 = Yes
+		* 1.1 = Abort
+		* 1.2 = Retry
+		* 1.3 = Ignore
+		* 1.4 = Cancel
+		* 1.5 = TryAgain
+		* 1.6 = Continue
+		* 1.7 = Help
+		* 1.8 = OK
+		* 1.9 = ClosedBox
+		* 2 = broken
 		*/
-		float ShowBox(LPCTSTR Message, LPCTSTR Caption, UINT Type);
+		double ShowBox(LPCTSTR Message, LPCTSTR Caption, UINT Type);
 
 
 };
